Rename package from skiba to forestspot
- Rename package directory, module skiba.py -> forestspot.py, and test files
- Update imports, pyproject metadata, mkdocs config, docs, README, CI, and issue templates
- Remove broken console_scripts entry point (skiba.cli never existed)
- Remove conda-forge badge (package is not published on conda-forge)

URLs under taraskiba/skiba and gskiba.streamlit.app are updated to the
post-rename targets where they refer to the repo; the Streamlit app URL is
left unchanged until the app itself is renamed.
