+++ docs/aggregated_point_extraction.md
@@ -1,4 +1,4 @@
 
 # aggregated_point_extraction module
 
-::: skiba.aggregated_point_extraction
+::: forestspot.aggregated_point_extraction
