+++ docs/buffer_and_sample.md
@@ -1,4 +1,4 @@
 
 # buffer_and_sample module
 
-::: skiba.buffer_and_sample
+::: forestspot.buffer_and_sample
