+++ docs/buffer_coordinates.md
@@ -1,4 +1,4 @@
 
 # buffer_coordinates module
 
-::: skiba.buffer_coordinates
+::: forestspot.buffer_coordinates
