+++ docs/buffer_method.md
@@ -1,3 +1,3 @@
 # buffer_method code
 
-::: skiba.buffer_method
+::: forestspot.buffer_method
