+++ docs/faq.md
@@ -1,5 +1,5 @@
 # FAQ
 
-Please see the streamlit app for FAQs ([streamlit](https://skibapython.streamlit.app/))
+Please see the streamlit app for FAQs ([streamlit](https://forestspotpython.streamlit.app/))
 
 Email me at: tskiba@vols.utk.edu[mailto:tskiba@vols.utk.edu]
