+++ docs/forestspot.md
@@ -1,0 +1,4 @@
+
+# forestspot module
+
+::: forestspot.forestspot
