+++ docs/installation.md
@@ -2,20 +2,20 @@
 
 ## Stable release
 
-To install skiba, run this command in your terminal:
+To install forestspot, run this command in your terminal:
 
 ```
-pip install skiba
+pip install forestspot
 ```
 
-This is the preferred method to install skiba, as it will always install the most recent stable release.
+This is the preferred method to install forestspot, as it will always install the most recent stable release.
 
 If you don't have [pip](https://pip.pypa.io) installed, this [Python installation guide](http://docs.python-guide.org/en/latest/starting/installation/) can guide you through the process.
 
 ## From sources
 
-To install skiba from sources, run this command in your terminal:
+To install forestspot from sources, run this command in your terminal:
 
 ```
-pip install git+https://github.com/taraskiba/skiba
+pip install git+https://github.com/taraskiba/forestspot
 ```
