+++ forestspot/__init__.py
@@ -1,4 +1,4 @@
-"""Top-level package for skiba."""
+"""Top-level package for forestspot."""
 
 __author__ = """Tara Skiba"""
 __email__ = "tskiba@vols.utk.edu"
