+++ forestspot/aggregated_point_extraction.py
